Made the coordinates and the velocity of the objects as floats. Also added the class code for time class.

+++ Engine/Entity.h
@@ -4,16 +4,16 @@
 class Entity
 {
 public: // Data Members
-	int x, y; // Top Left Coordinates of the object
-	int vx = 0, vy = 0; // Velocity of the object
+	float x, y; // Top Left Coordinates of the object
+	float vx = 0, vy = 0; // Velocity of the object
 	int r = 255, g = 255, b = 255; // Colour of the object
 	int width;
 
 protected: // Member Functions
 public:
 	Entity() = delete;
-	Entity(int x, int y, int width = 24);
-	Entity(int x, int y, int r, int g, int b, int width = 24);
+	Entity(float x, float y, int width = 24);
+	Entity(float x, float y, int r, int g, int b, int width = 24);
 
 	static bool checkCollision(const Entity& player, const Entity& poo);
 	virtual void keepInFrame(int right, int bottom);
